x64 and uefi

+++ kernel/serial.h
@@ -6,5 +6,5 @@ extern "C" {
 void serial_init();
 void serial_putc(char c);
 void serial_puts(const char *s);
-void serial_put_hex(uint32_t val);
+void serial_put_hex(uint64_t val);
 }
